Fix Tooltip when attaching to another mdl component
Use `Options.on` instead of `Html.Events.on` to attach
the event listener. This at least temporarily fixes the issue
where hovering over an MDL component such as `Button` with a tooltip
would cause the tooltip to stay visible.

NOTE: This might have unintended consequences when using `Tooltip`
with an MDL component with a ripple effect which also uses the
`mouseleave` event.